# Optimization: Reduce many array allocations in GetCustomAtttributes()'s inherited path