--- a/src/System.Private.CoreLib/src/System/Reflection/CustomAttribute.cs
+++ b/src/System.Private.CoreLib/src/System/Reflection/CustomAttribute.cs
@@ -1294,14 +1294,17 @@ namespace System.Reflection
                 type = type.BaseType as RuntimeType;
             }
 
-            if (result.Count == 0)
+            if (result.Count > 0)
             {
-                return useObjectArray ? Array.Empty<object>() : caType.GetEmptyArray();
+                object[] typedResult = CreateAttributeArrayHelper(arrayType, result.Count);
+                for (var i = 0; i < result.Count; i++)
+                {
+                    typedResult[i] = result[i];
+                }
+                return typedResult;
             }
-
-            object[] typedResult = CreateAttributeArrayHelper(arrayType, result.Count);
-            Array.Copy(result.ToArray(), 0, typedResult, 0, result.Count);
-            return typedResult;
+            // Return a cached, empty array
+            return useObjectArray ? Array.Empty<object>() : caType.GetEmptyArray();
         }
 
         internal static object[] GetCustomAttributes(RuntimeMethodInfo method, RuntimeType caType, bool inherit)
@@ -1342,9 +1345,17 @@ namespace System.Reflection
                 method = method.GetParentDefinition();
             }
 
-            object[] typedResult = CreateAttributeArrayHelper(arrayType, result.Count);
-            Array.Copy(result.ToArray(), 0, typedResult, 0, result.Count);
-            return typedResult;
+            if (result.Count > 0)
+            {
+                object[] typedResult = CreateAttributeArrayHelper(arrayType, result.Count);
+                for (var i = 0; i < result.Count; i++)
+                {
+                    typedResult[i] = result[i];
+                }
+                return typedResult;
+            }
+            // Return a cached, empty array
+            return useObjectArray ? Array.Empty<object>() : caType.GetEmptyArray();
         }
 
         internal static object[] GetCustomAttributes(RuntimeConstructorInfo ctor, RuntimeType caType)