--- a/src/System.Private.CoreLib/src/System/Reflection/CustomAttribute.cs
+++ b/src/System.Private.CoreLib/src/System/Reflection/CustomAttribute.cs
@@ -1279,14 +1279,14 @@ namespace System.Reflection
             RuntimeType.ListBuilder<object> result = new RuntimeType.ListBuilder<object>();
             bool mustBeInheritable = false;
             bool useObjectArray = (caType == null || caType.IsValueType || caType.ContainsGenericParameters);
-            Type arrayType = useObjectArray ? typeof(object) : caType;
+            RuntimeType arrayType = useObjectArray ? typeof(object) as RuntimeType : caType;
 
             while (pcaCount > 0)
                 result.Add(pca[--pcaCount]);
 
             while (type != (RuntimeType)typeof(object) && type != null)
             {
-                object[] attributes = GetCustomAttributes(type.GetRuntimeModule(), type.MetadataToken, 0, caType, mustBeInheritable, result);
+                object[] attributes = GetCustomAttributes(type.GetRuntimeModule(), type.MetadataToken, 0, caType, mustBeInheritable, ref result);
                 mustBeInheritable = true;
                 for (int i = 0; i < attributes.Length; i++)
                     result.Add(attributes[i]);
@@ -1294,17 +1294,12 @@ namespace System.Reflection
                 type = type.BaseType as RuntimeType;
             }
 
-            if (result.Count > 0)
+            object[] typedResult = CreateAttributeArrayHelper(arrayType, result.Count);
+            for (var i = 0; i < result.Count; i++)
             {
-                object[] typedResult = CreateAttributeArrayHelper(arrayType, result.Count);
-                for (var i = 0; i < result.Count; i++)
-                {
-                    typedResult[i] = result[i];
-                }
-                return typedResult;
+                typedResult[i] = result[i];
             }
-            // Return a cached, empty array
-            return useObjectArray ? Array.Empty<object>() : caType.GetEmptyArray();
+            return typedResult;
         }
 
         internal static object[] GetCustomAttributes(RuntimeMethodInfo method, RuntimeType caType, bool inherit)
@@ -1330,14 +1325,14 @@ namespace System.Reflection
             RuntimeType.ListBuilder<object> result = new RuntimeType.ListBuilder<object>();
             bool mustBeInheritable = false;
             bool useObjectArray = (caType == null || caType.IsValueType || caType.ContainsGenericParameters);
-            Type arrayType = useObjectArray ? typeof(object) : caType;
+            RuntimeType arrayType = useObjectArray ? typeof(object) as RuntimeType : caType;
 
             while (pcaCount > 0)
                 result.Add(pca[--pcaCount]);
 
             while (method != null)
             {
-                object[] attributes = GetCustomAttributes(method.GetRuntimeModule(), method.MetadataToken, 0, caType, mustBeInheritable, result);
+                object[] attributes = GetCustomAttributes(method.GetRuntimeModule(), method.MetadataToken, 0, caType, mustBeInheritable, ref result);
                 mustBeInheritable = true;
                 for (int i = 0; i < attributes.Length; i++)
                     result.Add(attributes[i]);
@@ -1345,17 +1340,12 @@ namespace System.Reflection
                 method = method.GetParentDefinition();
             }
 
-            if (result.Count > 0)
+            object[] typedResult = CreateAttributeArrayHelper(arrayType, result.Count);
+            for (var i = 0; i < result.Count; i++)
             {
-                object[] typedResult = CreateAttributeArrayHelper(arrayType, result.Count);
-                for (var i = 0; i < result.Count; i++)
-                {
-                    typedResult[i] = result[i];
-                }
-                return typedResult;
+                typedResult[i] = result[i];
             }
-            // Return a cached, empty array
-            return useObjectArray ? Array.Empty<object>() : caType.GetEmptyArray();
+            return typedResult;
         }
 
         internal static object[] GetCustomAttributes(RuntimeConstructorInfo ctor, RuntimeType caType)
@@ -1461,6 +1451,7 @@ namespace System.Reflection
                 RuntimeType attributeType;
                 IRuntimeMethodInfo ctor;
                 bool ctorHasParameters, isVarArg;
+                RuntimeType.ListBuilder<object> derivedAttributes = default;
 
                 // Optimization for the case where attributes decorate entities in the same assembly in which case 
                 // we can cache the successful APTCA check between the decorated and the declared assembly.
@@ -1471,7 +1462,7 @@ namespace System.Reflection
                     CustomAttributeRecord caRecord = car[i];
 
                     if (FilterCustomAttributeRecord(caRecord, scope, ref lastAptcaOkAssembly,
-                        decoratedModule, decoratedMetadataToken, attributeFilterType, mustBeInheritable, null, default,
+                        decoratedModule, decoratedMetadataToken, attributeFilterType, mustBeInheritable, null, ref derivedAttributes,
                         out attributeType, out ctor, out ctorHasParameters, out isVarArg))
                         return true;
                 }
@@ -1496,18 +1487,19 @@ namespace System.Reflection
         private static unsafe object[] GetCustomAttributes(
             RuntimeModule decoratedModule, int decoratedMetadataToken, int pcaCount, RuntimeType attributeFilterType)
         {
-            return GetCustomAttributes(decoratedModule, decoratedMetadataToken, pcaCount, attributeFilterType, false, default);
+            RuntimeType.ListBuilder<object> _ = default;
+            return GetCustomAttributes(decoratedModule, decoratedMetadataToken, pcaCount, attributeFilterType, false, ref _);
         }
 
         private static unsafe object[] GetCustomAttributes(
             RuntimeModule decoratedModule, int decoratedMetadataToken, int pcaCount,
-            RuntimeType attributeFilterType, bool mustBeInheritable, RuntimeType.ListBuilder<object> derivedAttributes)
+            RuntimeType attributeFilterType, bool mustBeInheritable, ref RuntimeType.ListBuilder<object> derivedAttributes)
         {
             MetadataImport scope = decoratedModule.MetadataImport;
             CustomAttributeRecord[] car = CustomAttributeData.GetCustomAttributeRecords(decoratedModule, decoratedMetadataToken);
 
             bool useObjectArray = (attributeFilterType == null || attributeFilterType.IsValueType || attributeFilterType.ContainsGenericParameters);
-            Type arrayType = useObjectArray ? typeof(object) : attributeFilterType;
+            RuntimeType arrayType = useObjectArray ? typeof(object) as RuntimeType : attributeFilterType;
 
             if (attributeFilterType == null && car.Length == 0)
                 return CreateAttributeArrayHelper(arrayType, 0);
@@ -1535,7 +1527,7 @@ namespace System.Reflection
 
                 if (!FilterCustomAttributeRecord(caRecord, scope, ref lastAptcaOkAssembly,
                                                  decoratedModule, decoratedMetadataToken, attributeFilterType, mustBeInheritable,
-                                                 attributes, derivedAttributes,
+                                                 attributes, ref derivedAttributes,
                                                  out attributeType, out ctor, out ctorHasParameters, out isVarArg))
                     continue;
 
@@ -1663,7 +1655,7 @@ namespace System.Reflection
             RuntimeType attributeFilterType,
             bool mustBeInheritable,
             object[] attributes,
-            RuntimeType.ListBuilder<object> derivedAttributes,
+            ref RuntimeType.ListBuilder<object> derivedAttributes,
             out RuntimeType attributeType,
             out IRuntimeMethodInfo ctor,
             out bool ctorHasParameters,
@@ -1687,7 +1679,7 @@ namespace System.Reflection
 
             // Ensure if attribute type must be inheritable that it is inhertiable
             // Ensure that to consider a duplicate attribute type AllowMultiple is true
-            if (!AttributeUsageCheck(attributeType, mustBeInheritable, derivedAttributes))
+            if (!AttributeUsageCheck(attributeType, mustBeInheritable, ref derivedAttributes))
                 return false;
 
             // Windows Runtime attributes aren't real types - they exist to be read as metadata only, and as such
@@ -1759,7 +1751,7 @@ namespace System.Reflection
 
         #region Private Static Methods
         private static bool AttributeUsageCheck(
-            RuntimeType attributeType, bool mustBeInheritable, RuntimeType.ListBuilder<object> derivedAttributes)
+            RuntimeType attributeType, bool mustBeInheritable, ref RuntimeType.ListBuilder<object> derivedAttributes)
         {
             AttributeUsageAttribute attributeUsageAttribute = null;
 
@@ -1859,8 +1851,14 @@ namespace System.Reflection
             blobStart = (IntPtr)pBlobStart;
         }
 
-        private static object[] CreateAttributeArrayHelper(Type elementType, int elementCount)
+        private static object[] CreateAttributeArrayHelper(RuntimeType elementType, int elementCount)
         {
+            // If we have 0 elements, don't allocate a new array
+            if (elementCount == 0)
+            {
+                return elementType.GetEmptyArray();
+            }
+
             return (object[])Array.UnsafeCreateInstance(elementType, elementCount);
         }
         #endregion