This is a follow-up to many allocations recognized in #20448.

A lot of methods ultimately call through `CustomAttributes.GetCustomAttributes()`. In the inherited search path (default for most searches above), the inheritance path of a class is traversed, resulting in an array allocation per crawled type, a copy to the overall List<object> and then after that - in current code - [an array allocation from that list and a typed array allocation it's copied to](https://github.com/dotnet/coreclr/blob/8b03dbeb90c81b8095b949b44553ff970e169851/src/System.Private.CoreLib/src/System/Reflection/CustomAttribute.cs#L1297-L1299).

However, in the common miss case as simple as:
```c#
typeof(T).GetCustomAttributes(typeof(TAttr), true);
```
...and many other overloads, all the same path underneath...

We can avoid the last 2 arrays in the miss case. We have the `List<object>` to go off of. If that's a zero-entry list, we can return an `Array.Empty<TAttr>()`. That's effectively what this change does.

While converting the entire attribute pipeline to generics is problematic and has issue since some `object[]` return abstracts aren't sealed, we can at least somewhat trivially cache an array per attribute type (only one static, ultimately from `Array.Empty<T>` underneath) and return that for the miss case.

There are far more wins to be had here, but they require more changes.

Benchmarks are in flux at the moment (see #20763), but the following benchmark:
```c#
[MemoryDiagnoser]
public class AttributePerf
{
    private static readonly MethodInfo AttributedOverride = typeof(AttributedClass).GetMethod(nameof(AttributedClass.FooAttributed));
    private static readonly MethodInfo NonAttributedOverride = typeof(NonAttributedClass).GetMethod(nameof(NonAttributedClass.FooAttributed));
    private static readonly MethodInfo AttributedBase = typeof(BaseClass).GetMethod(nameof(BaseClass.FooAttributed));
    private static readonly MethodInfo NonAttributedBase = typeof(BaseClass).GetMethod(nameof(BaseClass.FooUnattributed));

    [Benchmark(Description = "Class: Hit (inherit)")]
    public object[] ClassHitInherit() => typeof(AttributedClass).GetCustomAttributes(typeof(MyAttribute), true);
    [Benchmark(Description = "Class: Miss (inherit)")]
    public object[] ClassMissInherit() => typeof(NonAttributedClass).GetCustomAttributes(typeof(MyAttribute), true);

    [Benchmark(Description = "Class: Hit (no inherit)")]
    public object[] ClassHit() => typeof(AttributedClass).GetCustomAttributes(typeof(MyAttribute), false);
    [Benchmark(Description = "Class: Miss (no inherit)")]
    public object[] ClassMiss() => typeof(NonAttributedClass).GetCustomAttributes(typeof(MyAttribute), false);

    [Benchmark(Description = "Method Override: Hit (inherit)")]
    public object[] MethodOverrideHitInherit() => AttributedOverride.GetCustomAttributes(typeof(MyAttribute), true);
    [Benchmark(Description = "Method Override: Miss (inherit)")]
    public object[] MethodOverrideMissInherit() => NonAttributedOverride.GetCustomAttributes(typeof(MyAttribute), true);

    [Benchmark(Description = "Method Override: Hit (no inherit)")]
    public object[] MethodOverrideHit() => AttributedOverride.GetCustomAttributes(typeof(MyAttribute), false);
    [Benchmark(Description = "Method Override: Miss (no inherit)")]
    public object[] MethodOverrideMiss() => NonAttributedOverride.GetCustomAttributes(typeof(MyAttribute), false);

    [Benchmark(Description = "Method Base: Hit (inherit)")]
    public object[] MethodBaseHitInherit() => AttributedBase.GetCustomAttributes(typeof(MyAttribute), true);
    [Benchmark(Description = "Method Base: Miss (inherit)")]
    public object[] MethodBaseMissInherit() => NonAttributedBase.GetCustomAttributes(typeof(MyAttribute), true);

    [Benchmark(Description = "Method Base: Hit (no inherit)")]
    public object[] MethodBaseHit() => AttributedBase.GetCustomAttributes(typeof(MyAttribute), false);
    [Benchmark(Description = "Method Base: Miss (no inherit)")]
    public object[] MethodBaseMiss() => NonAttributedBase.GetCustomAttributes(typeof(MyAttribute), false);

    [My("Test")]
    public class AttributedClass : BaseClass
    {
        [My("FooAttributed")]
        public override string FooAttributed() => null;
    }

    public class NonAttributedClass : BaseClass
    {
        public override string FooUnattributed() => null;
    }

    public class BaseClass
    {
        [My("Foo")]
        public virtual string FooAttributed() => null;
        public virtual string FooUnattributed() => null;
    }

    [AttributeUsage(AttributeTargets.All, AllowMultiple = false)]
    public class MyAttribute : Attribute
    {
        private readonly string _name;
        public MyAttribute(string name) => _name = name;
    }
}
```
Yields the following gains with this change:

```
BenchmarkDotNet=v0.11.2, OS=Windows 10.0.17763.55 (1809/October2018Update/Redstone5)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview-009722
  [Host]     : .NET Core 3.0.0-preview1-27101-02 (CoreCLR 4.6.27029.02, CoreFX 4.6.27018.01), 64bit RyuJIT
  Job-TLDGVD : .NET Core ? (CoreCLR 4.6.27101.0, CoreFX 4.7.18.53005), 64bit RyuJIT
Runtime=Core  Toolchain=CoreRun
```
#### Before
|                               Method |       Mean |     Error |     StdDev |     Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------- |-----------:|----------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
|               'Class: Hit (inherit)' | 2,893.6 ns | 57.856 ns | 138.620 ns | 2,844.8 ns |      0.1793 |           - |           - |               568 B |
|              'Class: Miss (inherit)' | 1,062.9 ns | 21.184 ns |  42.307 ns | 1,040.4 ns |      0.0343 |           - |           - |               112 B |
|            'Class: Hit (no inherit)' | 1,897.2 ns | 28.440 ns |  26.603 ns | 1,890.7 ns |      0.1221 |           - |           - |               384 B |
|           'Class: Miss (no inherit)' |   376.5 ns |  7.597 ns |   9.878 ns |   373.5 ns |      0.0072 |           - |           - |                24 B |
|     'Method Override: Hit (inherit)' | 4,015.6 ns | 69.482 ns |  58.020 ns | 4,013.4 ns |      0.2747 |           - |           - |               872 B |
|    'Method Override: Miss (inherit)' | 2,470.7 ns | 48.886 ns | 120.835 ns | 2,433.7 ns |      0.1717 |           - |           - |               544 B |
|  'Method Override: Hit (no inherit)' | 1,885.3 ns | 37.046 ns |  45.496 ns | 1,867.3 ns |      0.1259 |           - |           - |               400 B |
| 'Method Override: Miss (no inherit)' | 1,882.6 ns | 36.321 ns |  45.935 ns | 1,868.7 ns |      0.1221 |           - |           - |               384 B |
|         'Method Base: Hit (inherit)' | 2,308.3 ns | 25.619 ns |  20.001 ns | 2,315.3 ns |      0.1717 |           - |           - |               544 B |
|        'Method Base: Miss (inherit)' |   719.4 ns | 14.435 ns |  32.582 ns |   703.6 ns |      0.0277 |           - |           - |                88 B |
|      'Method Base: Hit (no inherit)' | 1,825.9 ns | 11.479 ns |  10.738 ns | 1,828.2 ns |      0.1221 |           - |           - |               384 B |
|     'Method Base: Miss (no inherit)' |   384.8 ns |  7.642 ns |   7.506 ns |   383.3 ns |      0.0072 |           - |           - |                24 B |

#### Cached Array & loop copy
|                               Method |       Mean |     Error |    StdDev |     Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------- |-----------:|----------:|----------:|-----------:|------------:|------------:|------------:|--------------------:|
|               'Class: Hit (inherit)' | 2,657.1 ns | 59.834 ns | 79.877 ns | 2,615.9 ns |      0.1678 |           - |           - |               536 B |
|              'Class: Miss (inherit)' |   852.2 ns |  5.066 ns |  4.491 ns |   852.0 ns |      0.0277 |           - |           - |                88 B |
|            'Class: Hit (no inherit)' | 1,880.9 ns |  9.299 ns |  8.243 ns | 1,879.0 ns |      0.1221 |           - |           - |               384 B |
|           'Class: Miss (no inherit)' |   374.1 ns |  5.301 ns |  4.958 ns |   372.4 ns |      0.0067 |           - |           - |                24 B |
|     'Method Override: Hit (inherit)' | 3,902.4 ns | 32.201 ns | 30.121 ns | 3,899.1 ns |      0.2670 |           - |           - |               840 B |
|    'Method Override: Miss (inherit)' | 2,345.2 ns | 17.477 ns | 16.348 ns | 2,348.3 ns |      0.1602 |           - |           - |               512 B |
|  'Method Override: Hit (no inherit)' | 1,967.4 ns | 24.553 ns | 21.765 ns | 1,968.3 ns |      0.1259 |           - |           - |               400 B |
| 'Method Override: Miss (no inherit)' | 1,906.1 ns |  6.111 ns |  4.771 ns | 1,905.1 ns |      0.1221 |           - |           - |               384 B |
|         'Method Base: Hit (inherit)' | 2,361.2 ns | 28.074 ns | 24.887 ns | 2,357.8 ns |      0.1602 |           - |           - |               512 B |
|        'Method Base: Miss (inherit)' |   519.0 ns |  3.167 ns |  2.963 ns |   518.6 ns |      0.0200 |           - |           - |                64 B |
|      'Method Base: Hit (no inherit)' | 1,964.2 ns | 39.248 ns | 63.378 ns | 1,930.2 ns |      0.1221 |           - |           - |               384 B |
|     'Method Base: Miss (no inherit)' |   413.8 ns |  8.035 ns | 20.740 ns |   409.5 ns |      0.0072 |           - |           - |                24 B |

#### Above with `ref` `ListBuilder<T>` usage
|                               Method |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------- |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|               'Class: Hit (inherit)' | 2,262.8 ns | 44.668 ns | 56.490 ns |      0.1297 |           - |           - |               416 B |
|              'Class: Miss (inherit)' |   571.1 ns |  2.599 ns |  2.304 ns |           - |           - |           - |                   - |
|            'Class: Hit (no inherit)' | 1,735.9 ns | 18.678 ns | 17.471 ns |      0.1221 |           - |           - |               384 B |
|           'Class: Miss (no inherit)' |   247.7 ns |  5.057 ns |  5.823 ns |           - |           - |           - |                   - |
|     'Method Override: Hit (inherit)' | 3,383.7 ns | 26.707 ns | 24.982 ns |      0.2251 |           - |           - |               720 B |
|    'Method Override: Miss (inherit)' | 2,055.4 ns | 13.720 ns | 12.834 ns |      0.1297 |           - |           - |               416 B |
|  'Method Override: Hit (no inherit)' | 1,780.7 ns | 34.609 ns | 35.541 ns |      0.1259 |           - |           - |               400 B |
| 'Method Override: Miss (no inherit)' | 1,745.9 ns | 10.034 ns |  9.385 ns |      0.1221 |           - |           - |               384 B |
|         'Method Base: Hit (inherit)' | 2,053.4 ns | 14.508 ns | 13.571 ns |      0.1297 |           - |           - |               416 B |
|        'Method Base: Miss (inherit)' |   375.3 ns |  7.589 ns |  9.868 ns |           - |           - |           - |                   - |
|      'Method Base: Hit (no inherit)' | 1,780.4 ns | 39.946 ns | 54.679 ns |      0.1221 |           - |           - |               384 B |
|     'Method Base: Miss (no inherit)' |   254.0 ns |  1.739 ns |  1.541 ns |           - |           - |           - |                   - |

The gains are somewhat trivial on a single call (2 arrays and 24 bytes), but this is a path called a lot at scale and hopefully we agree multiplier on these changes is huge.

cc @benaadams @stephentoub @adamsitnik @noahfalk