@NickCraver 2018-11-03
@jkotas @stephentoub another round of fixes & tweaks in - numbers updated in the post, looking good but more is still on the table. I yanked in your latest changes here and cleaned up the cases a bit per-method - ready for a look when you have time. I'll poke a bit more at the remaining allocations here - there's still a lot of arrays to relay private methods results up and down but they're shared so much I'm not seeing a clear refactor path here yet (and such should probably be a different PR, unless you disagree).