@NickCraver 2018-11-03
Just FYI: I'm perplexed on how to eliminate the overall type-as-parameter and that being the return type problem here, but I'm poking `PseudoCustomAttribute` now. I think it can safely use `ListBuilder<T>` (we should probably move this...) underneath - will throw up another PR after this one lands if so.