@NickCraver 2018-11-04
I have a few more changes to further reduce allocations - since this is reviewed I'll fire up another PR after unless it's wanted here.

- [Eliminating PseudoCustomAttribute allocations](https://github.com/NickCraver/coreclr/commit/b91f9481c03e61f2789913a1ea1edcce2b5ac6d0)
- [Eliminating the inherited array allocation & copy](https://github.com/NickCraver/coreclr/commit/b5c3f4d7e55c2d5b4b495c8fca57022256ffe591)

Those two further reduce allocations on hits, more so in the inherited crawl path:

|                               Method |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------- |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|               'Class: Hit (inherit)' | 2,402.0 ns | 25.016 ns | 23.400 ns |      0.1221 |           - |           - |               384 B |
|              'Class: Miss (inherit)' |   578.8 ns | 12.380 ns | 10.338 ns |           - |           - |           - |                   - |
|            'Class: Hit (no inherit)' | 2,147.8 ns | 41.878 ns | 58.708 ns |      0.1221 |           - |           - |               384 B |
|           'Class: Miss (no inherit)' |   336.3 ns | 15.546 ns | 13.781 ns |           - |           - |           - |                   - |
|     'Method Override: Hit (inherit)' | 3,509.8 ns | 37.309 ns | 34.899 ns |      0.2060 |           - |           - |               656 B |
|    'Method Override: Miss (inherit)' | 2,253.3 ns | 68.342 ns | 78.703 ns |      0.1221 |           - |           - |               384 B |
|  'Method Override: Hit (no inherit)' | 2,142.0 ns | 19.804 ns | 18.525 ns |      0.1259 |           - |           - |               400 B |
| 'Method Override: Miss (no inherit)' | 2,151.8 ns | 20.700 ns | 19.362 ns |      0.1221 |           - |           - |               384 B |
|         'Method Base: Hit (inherit)' | 2,260.4 ns | 30.418 ns | 25.401 ns |      0.1221 |           - |           - |               384 B |
|        'Method Base: Miss (inherit)' |   367.3 ns |  3.426 ns |  3.037 ns |           - |           - |           - |                   - |
|      'Method Base: Hit (no inherit)' | 2,166.5 ns | 37.267 ns | 33.037 ns |      0.1221 |           - |           - |               384 B |
|     'Method Base: Miss (no inherit)' |   331.0 ns |  1.943 ns |  1.723 ns |           - |           - |           - |                   - |

`CustomAttributeData.GetCustomAttributeRecords` is the next item I believe, need to think about how to best eliminate that intermediary given the usually small-size arrays it creates. Related question: s there a `struct` friendly version of `ListBuilder<T>` that already exists?