@jkotas 2018-11-04
> have a few more changes to further reduce allocations

I have taken a cursory look and these look great. Thank you!

> Related question: s there a struct friendly version of `ListBuilder<T>` that already exists?

I am not sure what is struct unfriendly in `ListBuilder<T>`. I see that the `T` is constrained to `class`, but it should not be too hard to make it work for valuetypes as well.